legal-doc-simplifier/backend/requirements.txt
flask==2.3.3
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain==0.0.350
langchain-google-genai==0.0.6
pytesseract==0.3.10
Pillow==10.0.1
python-docx==0.8.11
pymongo==4.6.0
werkzeug==2.3.7
gunicorn==21.2.0
PyPDF2==3.0.1